reverted back to old tests
