+++ test/test_bpmn.py
@@ -1,4 +1,5 @@
 # type: ignore
+import inspect
 from unittest.mock import MagicMock
 from returns.pipeline import is_successful
 from bpmncwpverify.promela_gen_visitor import PromelaGenVisitor
@@ -62,20 +63,30 @@ def test_xml_parser():
     assert isinstance(gateway_12r266n, ParallelGatewayNode)
     assert gateway_12r266n is not None
     assert gateway_12r266n.name == "both"
+    assert len(gateway_12r266n.in_flows) == 2
     assert len(gateway_12r266n.out_flows) == 2
+    assert "Flow_08e7qxg" in {flow.id for flow in gateway_12r266n.in_flows}
+    assert "Flow_1wl740o" in {flow.id for flow in gateway_12r266n.in_flows}
     assert "Flow_1kx5xjv" in {flow.id for flow in gateway_12r266n.out_flows}
     assert "Flow_13xpfx7" in {flow.id for flow in gateway_12r266n.out_flows}
 
     gateway_0s1i42o = process["Gateway_0s1i42o"]
     assert gateway_0s1i42o is not None
     assert gateway_0s1i42o.name == "end both"
+    assert len(gateway_0s1i42o.in_flows) == 2
     assert len(gateway_0s1i42o.out_flows) == 1
+    assert "Flow_14s5onf" in {flow.id for flow in gateway_0s1i42o.in_flows}
+    assert "Flow_1oezfcg" in {flow.id for flow in gateway_0s1i42o.in_flows}
     assert "Flow_0feadgd" in {flow.id for flow in gateway_0s1i42o.out_flows}
 
     gateway_1pm4ghz = process["Gateway_1pm4ghz"]
     assert gateway_1pm4ghz is not None
     assert gateway_1pm4ghz.name == "payment and terms agreed"
+    assert len(gateway_1pm4ghz.in_flows) == 3
     assert len(gateway_1pm4ghz.out_flows) == 5
+    assert "Flow_0feadgd" in {flow.id for flow in gateway_1pm4ghz.in_flows}
+    assert "Flow_127sd29" in {flow.id for flow in gateway_1pm4ghz.in_flows}
+    assert "Flow_00oxr95" in {flow.id for flow in gateway_1pm4ghz.in_flows}
     assert "Flow_0yqye0v" in {flow.id for flow in gateway_1pm4ghz.out_flows}
     assert "Flow_0diuub0" in {flow.id for flow in gateway_1pm4ghz.out_flows}
     assert "Flow_0q6dz8p" in {flow.id for flow in gateway_1pm4ghz.out_flows}
@@ -85,55 +96,74 @@ def test_xml_parser():
     gateway_000lymc = process["Gateway_000lymc"]
     assert gateway_000lymc is not None
     assert gateway_000lymc.name == "both1"
+    assert len(gateway_000lymc.in_flows) == 1
     assert len(gateway_000lymc.out_flows) == 2
+    assert "Flow_0ct87dl" in {flow.id for flow in gateway_000lymc.in_flows}
     assert "Flow_0jmvvxc" in {flow.id for flow in gateway_000lymc.out_flows}
     assert "Flow_1y66pph" in {flow.id for flow in gateway_000lymc.out_flows}
 
     gateway_0cy7rs7 = process["Gateway_0cy7rs7"]
     assert gateway_0cy7rs7 is not None
     assert gateway_0cy7rs7.name == "end both1"
+    assert len(gateway_0cy7rs7.in_flows) == 2
     assert len(gateway_0cy7rs7.out_flows) == 1
+    assert "Flow_1sx1rdt" in {flow.id for flow in gateway_0cy7rs7.in_flows}
+    assert "Flow_0znbe82" in {flow.id for flow in gateway_0cy7rs7.in_flows}
     assert "Flow_1cm84v3" in {flow.id for flow in gateway_0cy7rs7.out_flows}
 
     # Activities
     activity_1unsjkg = process["Activity_1unsjkg"]
     assert activity_1unsjkg is not None
     assert activity_1unsjkg.name == "2-Buyer and Seller negotiate terms"
+    assert len(activity_1unsjkg.in_flows) == 1
     assert len(activity_1unsjkg.out_flows) == 1
+    assert "Flow_1kx5xjv" in {flow.id for flow in activity_1unsjkg.in_flows}
     assert "Flow_1oezfcg" in {flow.id for flow in activity_1unsjkg.out_flows}
 
     activity_1t579ox = process["Activity_1t579ox"]
     assert activity_1t579ox is not None
     assert activity_1t579ox.name == "3-Buyer and Seller negotiate price"
+    assert len(activity_1t579ox.in_flows) == 1
     assert len(activity_1t579ox.out_flows) == 1
+    assert "Flow_13xpfx7" in {flow.id for flow in activity_1t579ox.in_flows}
     assert "Flow_14s5onf" in {flow.id for flow in activity_1t579ox.out_flows}
 
     activity_1qm7qck = process["Activity_1qm7qck"]
     assert activity_1qm7qck is not None
     assert activity_1qm7qck.name == "1-Buyer sees desired backpack and meets Seller"
+    assert len(activity_1qm7qck.in_flows) == 1
     assert len(activity_1qm7qck.out_flows) == 1
+    assert "Flow_0oiwgjd" in {flow.id for flow in activity_1qm7qck.in_flows}
     assert "Flow_1wl740o" in {flow.id for flow in activity_1qm7qck.out_flows}
 
     activity_1qqx4aq = process["Activity_1qqx4aq"]
     assert activity_1qqx4aq is not None
     assert activity_1qqx4aq.name == "7a-Buyer hands cash payment to Seller"
+    assert len(activity_1qqx4aq.in_flows) == 1
     assert len(activity_1qqx4aq.out_flows) == 1
+    assert "Flow_0jmvvxc" in {flow.id for flow in activity_1qqx4aq.in_flows}
     assert "Flow_0znbe82" in {flow.id for flow in activity_1qqx4aq.out_flows}
 
     activity_1rfm4sh = process["Activity_1rfm4sh"]
     assert activity_1rfm4sh is not None
     assert activity_1rfm4sh.name == "7b-Seller hands Backpack to Buyer"
+    assert len(activity_1rfm4sh.in_flows) == 1
     assert len(activity_1rfm4sh.out_flows) == 1
+    assert "Flow_1y66pph" in {flow.id for flow in activity_1rfm4sh.in_flows}
     assert "Flow_1sx1rdt" in {flow.id for flow in activity_1rfm4sh.out_flows}
 
     # End event
     end_event_0wqympo = process["Event_0wqympo"]
     assert end_event_0wqympo is not None
     assert end_event_0wqympo.name == "Purchase Failed"
+    assert len(end_event_0wqympo.in_flows) == 1
+    assert "Flow_0diuub0" in {flow.id for flow in end_event_0wqympo.in_flows}
 
     end_event_1y6wxsp = process["Event_1y6wxsp"]
     assert end_event_1y6wxsp is not None
     assert end_event_1y6wxsp.name == "Purchase Completed"
+    assert len(end_event_1y6wxsp.in_flows) == 1
+    assert "Flow_1cm84v3" in {flow.id for flow in end_event_1y6wxsp.in_flows}
 
     # Start event
     start_event_1y8wbre = process["StartEvent_1y8wbre"]
@@ -141,6 +171,7 @@ def test_xml_parser():
     assert start_event_1y8wbre is not None
     assert start_event_1y8wbre.name == "Start7"
     assert len(start_event_1y8wbre.out_flows) == 1
+    assert len(start_event_1y8wbre.in_flows) == 0
     assert "Flow_0oiwgjd" in {flow.id for flow in start_event_1y8wbre.out_flows}
 
     # Flows
@@ -151,6 +182,16 @@ def test_xml_parser():
 def test_flow_traversal():
     bpmn: Bpmn = Bpmn.from_xml(workflow_bpmn_path)
     visitor = PromelaGenVisitor()
+
+    for name, method in inspect.getmembers(visitor, predicate=inspect.ismethod):
+        signature = inspect.signature(method)
+
+        return_annotation = signature.return_annotation
+
+        if return_annotation is bool:
+            setattr(visitor, name, MagicMock(return_value=True))
+        else:
+            setattr(visitor, name, MagicMock(return_value=None))
 
     assert is_successful(bpmn)
 
@@ -169,3 +210,23 @@ def test_flow_traversal():
 
     for flow in bpmn.processes[0].flows.values():
         assert flow.accept.call_count == EXPECTED_CALL_COUNT
+
+
+def test_graph_viz_generation():
+    bpmn: Bpmn = Bpmn.from_xml(workflow_bpmn_path)
+
+    assert is_successful(bpmn)
+
+    bpmn = bpmn.unwrap()
+
+    bpmn.generate_graph_viz()
+
+
+def test_promela_generation():
+    bpmn: Bpmn = Bpmn.from_xml(workflow_bpmn_path)
+
+    assert is_successful(bpmn)
+
+    bpmn = bpmn.unwrap()
+
+    bpmn.generate_promela()
